Checkpoint native numeric carrier and provider corrections

+++ crates/tsonic_rust_js/src/array/entries.rs
@@ -28,7 +28,7 @@ impl<T> JsArrayEntries<T> {
 }
 
 impl<T: Clone> JsArrayEntries<T> {
-    pub fn next_result(&self) -> IteratorResult<(f64, T), Undefined> {
+    pub fn next_result(&self) -> IteratorResult<(usize, T), Undefined> {
         match self.clone().next() {
             Some(value) => IteratorResult::yielded(value),
             None => IteratorResult::completed(Undefined),
@@ -52,7 +52,7 @@ impl<T> ObjectIdentityCarrier for JsArrayEntries<T> {
 }
 
 impl<T: Clone> Iterator for JsArrayEntries<T> {
-    type Item = (f64, T);
+    type Item = (usize, T);
 
     fn next(&mut self) -> Option<Self::Item> {
         let mut state = self.state.borrow_mut();
@@ -64,7 +64,7 @@ impl<T: Clone> Iterator for JsArrayEntries<T> {
         let value = array.get(state.index)?;
         let index = state.index;
         state.index += 1;
-        Some((index as f64, value))
+        Some((index, value))
     }
 }
